Client error handling as first-loaded separate script
